Forcing Tone to Nearest Basis Function
Forcing the radians per sample to be "dead nuts" on a basis function is the absolute best way to test purity. By going "dead nuts", there is zero appreciable leakage and thereby no need to apply a windowing function. This is absolutely the best spectral purity test possible for the flying phasor tone generator.